walette-holder/pyproject.toml
[tool.poetry]
name = "walette"
version = "0.1.0"
description = ""
authors = ["Joshua Thijssen <jthijssen@noxlogic.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.13"
uvicorn = "^0.34.3"
python-jose = "^3.5.0"
cryptography = "^45.0.4"
psycopg = "^3.2.9"
jwcrypto = "^1.5.6"
httpx = "^0.28.1"
jinja2 = "^3.1.6"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
